pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amazing-deck"
version = "0.1.0"
description = "Analyze PowerPoint templates and generate on-brand decks with reusable recipes (charts, timelines, KPI cards, asks)."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "amazing-deck contributors"}]
keywords = ["powerpoint", "pptx", "template", "deck", "presentation", "brand"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business",
]
dependencies = [
    "python-pptx>=0.6.23",
    "Pillow>=10.0.0",
]

[project.scripts]
amazing-deck = "amazing_deck.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
